requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-eventhub==5.11.4
azure-identity==1.15.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
retry==0.9.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0 